Move Mapping to own file and other cleanup
